A new 'reindex' function, to derive 'mapi' from 'map'.
  val reindex : (int -> 'a -> 'b) -> ('a -> 'b)
Does the right thing so that 'mapi f' = 'map (reindex f)'. It uses
a side-effect on a counter to add integer index starting from 0.

It will be used to add a 'mapi' function to all data structure
exposing a 'map' function : if the data structure is not naturally
associative (in which case 'mapi' use the keys as additionnal
argument), 'mapi' indexes over 'int' as permitted by the `reindex`
function.

I have put `reindex` into batInterfaces for now, as it is used for
interface harmonization, but it may move to a better place later.