Simplification of the (Mono)Mappable((Mono)Assoc) interfaces.
After discussion on the mailing-list, I decided to simplify the
Mappable interfaces. By assuming that all modules implement the 'mapi'
interface, we can reduce the number of different interfaces from 5 to
3 : we only have MappableAssoc, MappableMonoAssoc and
MonoMappableMonoAssoc left. I renamed them respectively Mappable2,
Mappable1 and Mappable0, using the fact that the type `'a mappable`
has 2, 1 or 0 polymorphic parameters. Mappable is an alias for
Mappable1, both because it's maybe the most natural/common signature,
and for compatibility reasons.