factorized implementation of functorized Map and PMap
The implementation of PMap and batteries-added functorized Map
functions were factorized under a common "concrete" implementation,
directly handling the balanced tree data structure.

We previously had three kind of (P)Map functions :

- standard stdlib's Map functions included from the depended-upon
  stdlib's Map module

- PMap functions, implemented in batteries batMap.ml

- additionnal function to enhance the functorized Map module, that are
  implemented on top of the included Map module by breaking the
  balanced-tree encapsulation to manipulate it directly

When adding a new function to PMap, the correspondent Map function was
shortly asked for, and reimplemented in the Map way (constant keys and
comparison function). This quickly led to a copious amount of code
duplication between Map and PMap.

The new design use a single implementation, Concrete (which directly
exposes the data structure and is not published to the users), from
which we derive the specialized Map and PMap implementations.

We currently still rely on stdlib's Map module (only the "additional
Map functions" are implemented using Concrete), but we could
definitely switch to Concrete and remove the dependency on stdlib's
Map and the somewhat-hackish data encapsulation breaking.