dead code removal in batMap.ml
I assume this commented implementation was a remain of a previous
implementation of the "enum" implementation. The main difference is
that the current one returns the elements in (increasing) sorted
order, instead of returning the higher elements first before
enumerating the subtrees. I think the reified data structure ('a iter)
(which seems inspired form stdlib's) is used to gain tail-recursivity
(it's a reification of a potential call stack).