reintroduction of Map.split function
[split] is a mandatory addition, as it has been added to stdlib 3.12 :
if we want to keep the stdlib->batteries compatibility for 3.12, we
need to add split, as well as the following missing functions :
  singleton, cardinal, for_all, exists, filter, bindings, choose

The implementation is similar to the one found in patch
2098d915c93e815cb8767c7bf9774b3dd39d59da Anton Novikov, but is inside
the new Concrete module.

Remark: `split` has not yet been added to batSplay.ml; in order not to
break its batMap.S interface, I added a dummy `split` raising an
error. An implementation should be provided soon.