Fix for the evaluation order of BatMap.{map,mapi}
In extlib's PMap, the evaluation order of [map] and [mapi] is
unspecified. However, the least surprise principle recommends to still
implement the most natural order when possible; in that case, it is
the "increasing key order", which is already described in the
documentation of stdlib's Map module. Changing extlib's implementation
to respect this order was simply a matter of explicit let..in
introduction, with no performance cost.
I have not changed the documentation which stills describe an
"unspecified order". Wheter we document the implemented order or not
is an independent question.

When writing test, I also rediscovered a bug in stdlib's
Map.S.{map,mapi} (3.11, fixed in 3.12), which also doesn't respect the
(specified) increasing key order
( http://caml.inria.fr/mantis/view.php?id=4012 ). This was fixed by
replacing stdlib's function by the Concrete implementation.