a new benchsuite/test_int file to benchmark the BatInt module
During experimentation on BatMap performances, I noticed that the
choice of the BatInt.compare function can have a small but noticeable
impact on BatMap performances. This new test helps choosing a fast
BatInt.compare function.

In the future it would be interesting to also measure other functions
performance, for example to evaluate the concrete cost of the
overflow-safe operation variants.