a fix to the benchsuite/test_maps module : ensure that the same comparison function is used in PMap and Map benchmarks.
previous performance results for test_maps were a bit strange in that,
once the balancing issue was solved, the previously slower PMap tests
turned out to be *faster* than functorized Map, instead of having
similar performances.

I discovered that the important difference between the two module is
that the functorized Map used the comparison function from BatInt,
while PMap uses Pervasives.compare by default. It turns out that
Pervasives.compare was significantly faster, and that impacted (P)Map
tests.

The tests are now fixed so that the same comparison function
(Pervasives.compare) is used everywhere. As expected, performances of
Map and PMap are now exactly identical.