The Map.merge function ported from stdlib 3.12
The merge function implementation from stdlib is written in the
functorized Map setting, assuming that all maps share a common
comparison function.
This assumption is not true in general in the PMap setting. A new
implementation of merge was designed, that does not rely on that
invariant, in the style of [PMap.union].