improved implementation of Array.filteri, bringing upto 100% performances improvements for fast predicates
Here are the performances results on my machine on various tests
cases. Here is how each test case should be read : `(p/q) for N` means
that the filter predicate keeps `p` out of `q` elements in an array of
input size `N`.

All tests performed using the ocaml-benchmark library :
  http://ocaml-benchmark.forge.ocamlcore.org/

For computations that took less than 1 second, the rate per second is
reported, otherwise the time for one iteration. In all cases, the
relative gain from new_impl (the new implementation) over bitset (the
old implementation) is what we should look at.

The new implementation shines in particular on small inputs (because
it uses lists), or when the result is small wrt. the input array
(because bitset implementation did two traversals of the input array,
instead of one of the input and one on the output).

Those tests of course only make sense for small predicates. Otherwise
the efficiency of filter is drowned into the computation time for the
predicate. But fast predicate are often used in practice (think of
"not None"), so it's reasonable to optimize for them if it doesn't
degrade performances overall.

Rate (9/10 for 100)
    bitset  5491/s       --       -53%
  new_impl 11793/s       115%     --

Rate (1/2 for 100)
    bitset  7190/s       --       -53%
  new_impl 15326/s       113%     --

Rate (1/20 for 100)
    bitset 11441/s       --       -50%
  new_impl 22883/s       100%     --

Rate (9/10 for 50000)
    bitset 9.77/s        --       -0%
  new_impl 9.80/s        0%       --

Rate (1/2 for 50000)
    bitset 13.2/s        --       -18%
  new_impl 16.1/s        22%      --

Rate (1/20 for 50000)
    bitset 21.5/s        --       -48%
  new_impl 41.2/s        92%      --

s/iter (9/10 for 1000000)
    bitset 2.14          --       -13%
  new_impl 1.86          15%      --

s/iter (1/2 for 1000000)
    bitset 1.61          --       -25%
  new_impl 1.21          33%      --

Rate (1/20 for 1000000)
    bitset 1.07/s        --       -47%
  new_impl 2.03/s        90%      --

Rate (1/1000 for 1000000)
    bitset 1.16/s        --       -50%
  new_impl 2.34/s        101%     --