A testsuite for the batUref Union-Find module
When writing the testsuite, I found a bug in the 'unite'
implementation. The test against it is currently disabled -- I try to
get clean states on commits to facilitate bisects.

The printing functions are not tested. They're not part of the core
logic and I'm not sure what to check.