batUref: allocate less by having head metadata in a mutable struct
By having heads being mutable structure, we can write things such as
   match !ur with
     | Ranked head -> head.data <- foo
instead of
   match !ur with
     | Ranked (x, r) -> ur := Ranked (foo, r)
which performed one Ranked allocation.

The price to pay is that Ranked now point to a struct instead of
having its metadata packed with the constructor. It should not be an
issue as, especially under unification-heavy scenario, there are more
Ptr nodes than heads.