refactor check_measures to avoid toplevel shadowing
This redefinition of "check_measures" was spotted by
'notfancy'. I moved the recursive helper function inside the outside
function instead of naming it check_measures_aux because this '_aux'
pattern is consistently used where the '_aux' function is called on
(('a, 'm) node, 'm) fg trees, which is not the case here.