move some code from qtest's hardcoded preamble to qtest-preamble.ml
I moved Random.self_init() out of the hardcoded preamble because
I believe qtest user want to be in control of Random's
reproductibility -- I understand Quickcheck is more useful with an
initialized RNG. Maybe we should engineer Quickcheck to use
a separate, globally accessible and initializable seed.

I kept the alias for Q in the hardcoded preamble because it is
advertized in the documentation. My personal opinion would be to use
as few "magic" as possible; for example, disable the hardcoded
preamble in case a preamble is given by the user -- so that qtest has
useful aliases by default, but the user can easily drop them if
they're not useful.