a better fix for the BatEnum.{from,push} interaction bug
This commit actually reverts the previous fix, and implement a quite
different one. I observed a small performance degradation of
BatEnum.from on a micro-benchmark, but the effects should be
neglectible on real code (anything doing more work for next-element
production than `BatEnum.from (fun () -> ())` ), and the fix is more
robust.